Cleanup and refactor demo data and chart components
- Move demo data to a new file `demo_data_settings.ts`
- Remove console logs from various components
- Simplify FacetContainer and PlotChartPanel rendering logic
- Remove unnecessary imports and debug statements
- Minor type and provider adjustments
